+++ src/editing/tools/Move.js
@@ -3,7 +3,7 @@ import Geometry from '../../math3d/Geometry.js';
 import SculptBase from './SculptBase.js';
 
 class Move extends SculptBase {
-
+  // v0.7.446: Cache Bust
   constructor(main) {
     super(main);