+++ src/mesh/meshStatic/MeshStatic.js
@@ -4,6 +4,7 @@ import MeshData from '../MeshData.js';
 import RenderData from '../RenderData.js';
 import Enums from '../../misc/Enums.js';
 
+// v0.7.457: Force Reload
 class MeshStatic extends Mesh {
 
   constructor(gl) {